feat: ChatVM의 NestType이던 OpenAIRecommendation를 외부로 분리 및 접근제어를 internal로 수정
